basic seperation
