+++ Client/src/components/notes/ToolbarButton.jsx
@@ -1,0 +1,25 @@
+const ToolbarButton = ({ onClick, isActive, icon, title }) => (
+  <button
+    onClick={onClick}
+    title={title}
+    className="p-1.5 border-none cursor-pointer flex items-center justify-center transition-all"
+    style={{
+      backgroundColor: isActive ? "var(--btn)" : "transparent",
+      color: isActive ? "white" : "var(--txt)",
+    }}
+    onMouseEnter={(e) => {
+      if (!isActive) {
+        e.target.style.backgroundColor = "var(--bg-primary)";
+      }
+    }}
+    onMouseLeave={(e) => {
+      if (!isActive) {
+        e.target.style.backgroundColor = "transparent";
+      }
+    }}
+  >
+    {icon}
+  </button>
+);
+
+export default ToolbarButton;
